+++ lib/DS1302_Util/DS1302_Util.h
@@ -1,0 +1,10 @@
+#include <Arduino.h>
+#include <ThreeWire.h>
+#include <RtcDS1302.h>
+
+//Prototipo RTC
+void printDateTime(const RtcDateTime &dt);
+
+void init_RTC(void);
+
+void printTime(void);
